# docs/swagger.yaml
definitions:
...
  github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-array_internal_program_Program:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_program.Program'
        type: array
      message:
        type: string
    type: object
  github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-array_internal_smtp_Instance:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_smtp.Instance'
        type: array
      message:
        type: string
    type: object
...
  github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string:
    properties:
      data:
        $ref: '#/definitions/map_string_string'
      message:
        type: string
    type: object
...
  internal_invite.createInviteInput:
    properties:
      expiresAt:
        type: string
    type: object
  internal_invite.selfRegisterInput:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      studentId:
        type: string
    required:
    - studentId
    type: object
...
  internal_program.Program:
    properties:
      active:
        type: boolean
      description:
        type: string
      id:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  internal_program.createCourseInput:
    properties:
      active:
        type: boolean
      campus_id:
        type: string
      description:
        type: string
      name:
        type: string
    required:
    - active
    - campus_id
    - description
    - name
    type: object
  internal_program.updateCourseInput:
    properties:
      active:
        type: boolean
      description:
        type: string
      name:
        type: string
    type: object
  internal_smtp.Instance:
    properties:
      email:
        type: string
      host:
        type: string
      id:
        type: string
      iv:
        type: string
      password:
        type: string
      port:
        type: integer
    required:
    - email
    - host
    - iv
    - password
    - port
    type: object
  internal_smtp.createInstanceInput:
    properties:
      email:
        type: string
      host:
        type: string
      jwe:
        type: string
      password:
        type: string
      port:
        type: integer
    required:
    - email
    - host
    - jwe
    - password
    - port
    type: object
...
  map_string_string:
    additionalProperties:
      type: string
    type: object
info:
  contact: {}
paths:
...
  /invite/{courseId}:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      - description: Expiração opcional
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_invite.createInviteInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
      summary: Cria um convite para um curso
      tags:
      - invite
  /invite/self-register/{code}:
    post:
      consumes:
      - application/json
      parameters:
      - description: Código do convite
        in: path
        name: code
        required: true
        type: string
      - description: Dados do aluno
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_invite.selfRegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
      summary: Auto-registro de aluno via convite
      tags:
      - invite
...
  /program:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Dados do programa
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_program.createCourseInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
      summary: Cria um programa
      tags:
      - program
  /program/{id}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Program ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
      summary: Deleta um programa
      tags:
      - program
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Campus ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-array_internal_program_Program'
      summary: Lista programas por campus
      tags:
      - program
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Program ID
        in: path
        name: id
        required: true
        type: string
      - description: Campos para atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_program.updateCourseInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
      summary: Atualiza um programa
      tags:
      - program
  /smtp/instance:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-array_internal_smtp_Instance'
      summary: Lista instâncias SMTP do usuário
      tags:
      - smtp
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Dados SMTP
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_smtp.createInstanceInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
      summary: Cria uma instância SMTP
      tags:
      - smtp
...
swagger: "2.0"